backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
python-multipart==0.0.6
python-dotenv==1.0.0
pdfplumber==0.10.3
openai==1.3.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-magic==0.4.27
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9